samples/harp/src/python/requirements.txt
PyNaCl>=1.5.0
cryptography>=42.0.0
canonicaljson>=2.0.0
